requirements.txt
numpy==1.24.3
opencv-contrib-python==4.9.0.80
python-dotenv==1.0.1
